moved all gui elements to their own python scripts
